test(cli): consolidate test service coverage
